Decided, for diffing purposes, to move the generate-nodes stuff back to source-nodes for now
